chore: sepereate laborRightsApi
